repo organization and moving of files, readme created
