+++ analyses/WindRose/README.md
@@ -1,0 +1,26 @@
+# This folder is for analyses on data contamination sources using wind rose plots.
+
+The met data used here is courtesy of NOAA ESRL GMD. See the citations below. ftp://ftp.cmdl.noaa.gov/met/sum/README
+
+Mefford, T.K., M. Bieniulis, B. Halter, and J. Peterson,
+Meteorological Measurements, in CMDL Summary Report 1994 - 1995,
+No. 23, 1996, pg. 17.
+
+Herbert, G., M. Bieniulis, T. Mefford, and K. Thaut, Acquisition
+and Data Management Division, in CMDL Summary Report 1993,
+No. 22, 1994, pg. 57
+
+Herbert, G.A., J. Harris, M. Bieniulis, and J. McCutcheon,
+Acquisition and Data Management, in CMDL Summary Report
+1989, No. 18, 1990, Pg. 50.
+
+Herbert, G.A., E.R. Green, G.L. Koenig, and K.W. Thaut,
+Monitoring instrumentation for the continuous measurement
+and quality assurance of meteorological observations, NOAA
+Tech. Memo. ERL ARL-148, 44 pp, 1986.
+
+Herbert, G.A., E.R. Green, J.M. Harris, G.L. Koenig, S.J.
+Roughton, and K.W. Thaut, Control and Monitoring
+Instrumentation for the Continuous Measurement of
+Atmospheric CO2 and Meteorological Variables, J. of Atmos.
+and Oceanic Tech., 3, 414-421, 1986.
